SOURCE BETA - version 0.8 beta 1

+++ Sources/Beta Code/libraries/ArduinoJson/src/ArduinoJson/Polyfills/safe_strcmp.hpp
@@ -6,19 +6,27 @@
 
 #include <ArduinoJson/Namespace.hpp>
 
+#include <stdint.h>  // int8_t
+
 namespace ARDUINOJSON_NAMESPACE {
 
-inline int8_t safe_strcmp(const char* a, const char* b) {
-  if (a == b) return 0;
-  if (!a) return -1;
-  if (!b) return 1;
-  return static_cast<int8_t>(strcmp(a, b));
+inline int safe_strcmp(const char* a, const char* b) {
+  if (a == b)
+    return 0;
+  if (!a)
+    return -1;
+  if (!b)
+    return 1;
+  return strcmp(a, b);
 }
 
-inline int8_t safe_strncmp(const char* a, const char* b, size_t n) {
-  if (a == b) return 0;
-  if (!a) return -1;
-  if (!b) return 1;
-  return static_cast<int8_t>(strncmp(a, b, n));
+inline int safe_strncmp(const char* a, const char* b, size_t n) {
+  if (a == b)
+    return 0;
+  if (!a)
+    return -1;
+  if (!b)
+    return 1;
+  return strncmp(a, b, n);
 }
 }  // namespace ARDUINOJSON_NAMESPACE
